added logout
