+++ client/src/App.js
@@ -7,14 +7,15 @@ import {
 } from "react-router-dom";
 import "./App.css"
 import axios from 'axios';
-import Nav from './comp/Nav/Nav'
-import Footer from './comp/Footer/Footer'
-import Home from './pages/Home'
-import Login from './pages/Login/Login'
-import Register from './pages/Register/Register'
-import Weather from './pages/Weather'
-import Checklist from './pages/checklist/CheckList'
-import {useAuth, actions} from "./utils/authState"
+import Nav from './comp/Nav/Nav';
+import Footer from './comp/Footer/Footer';
+import Home from './pages/Home';
+import Login from './pages/Login/Login';
+import Register from './pages/Register/Register';
+import Weather from './pages/Weather';
+import Logout from './pages/Logout/logout';
+import Checklist from './pages/checklist/CheckList';
+import {useAuth, actions} from './utils/authState';
 
 export default function App() {
     const [authState, authDispatch] = useAuth();
@@ -43,6 +44,9 @@ export default function App() {
             ?<Login />
             :<Redirect to={"/"} />
           }
+        </Route>
+        <Route path="/logout" >
+          <Logout />
         </Route>
         <Route path="/register">
           <Register />
