+++ client/src/pages/Login/Login.jsx
@@ -7,7 +7,6 @@ import './Login.css'
 function Login() {
   const [loginUsername, setLoginUsername] = useState("");
   const [loginPassword, setLoginPassword] = useState("");
-  const [data, setData] = useState(null);
 
   const login = () => {
     axios({
@@ -21,27 +20,6 @@ function Login() {
     })
   };
 
-  const logout = () => {
-    axios({
-      method:"GET",
-      data: {
-        username: loginUsername,
-        password: loginPassword,
-      },
-      withCredentials: false,
-      url: "/api/logout",
-    })
-  };
-  // const getUser = () => {
-  //   Axios({
-  //     method: "GET",
-  //     withCredentials: true,
-  //     url: "http://localhost:3001/user",
-  //   }).then((res) => {
-  //     setData(res.data);
-  //     console.log(res.data);
-  //   });
-  // };
   return (
     <div className="loginBody">
       <Row>
@@ -58,18 +36,8 @@ function Login() {
               onChange={(e) => setLoginPassword(e.target.value)}
             />
             <Button className="button" onClick= {login}>Login</Button>
-          <Button className="button" onClick= {logout}>Logout</Button>
           </Card>
       </Row>
-      {/* <Row>
-        <Container>
-          <Card className="card" body inverse style={{ backgroundColor: '#f1f5', borderColor: '#f1f5' }}>
-            <CardTitle className="font" >Get User</CardTitle>
-            <Button className="button" onClick={getUser}>Submit</Button>
-            {data ? <h1>Welcome Back: {data.username}</h1> : null}
-          </Card>
-        </Container>
-      </Row> */}
     </div>
   );
 }
